add request_time to event model

+++ backend/api/events/schemas.py
@@ -15,7 +15,7 @@ class Event(BaseModel):
     user_id: Optional[int]
     token: Optional[str]
     ip: Optional[str]
-    request_time: Optional[int]
+    request_time: Optional[float]
     promoimage: Optional[str]
     promodata: Optional[dict]
